move app functionality to home component
