+++ frontend/src/app/app-routing.module.ts
@@ -1,11 +1,17 @@
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import { RouterModule, Routes } from '@angular/router';
+import { HomeComponent } from '@app/pages/home/home.component';
 
+const routes: Routes = [
+  { path: '', pathMatch: 'full', component: HomeComponent }, // fully match an empty route
 
-const routes: Routes = [];
+  // Redirect all to HomeComponent
+  { path: '**', redirectTo: '' }
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule]
 })
-export class AppRoutingModule { }
+export class AppRoutingModule {
+}
