+++ frontend/src/app/pages/home/home.component.html
@@ -1,0 +1,188 @@
+<mat-toolbar color="primary">
+  <span>Dotazník před očkováním proti nemoci COVID-19</span>
+</mat-toolbar>
+
+<main>
+  <mat-horizontal-stepper [linear]="true">
+    <mat-step *ngIf="basicInfoForm" [completed]="canProceedToStep2" label="Osobní informace">
+      <div class="container">
+        <mat-card>
+          <mat-card-content>
+            <form [formGroup]="basicInfoForm">
+              <mat-form-field>
+                <mat-label>Jméno</mat-label>
+                <input [(ngModel)]="patientInfo.firstName" formControlName="firstName" matInput placeholder="Vaše jméno" required>
+              </mat-form-field>
+
+              <mat-form-field>
+                <mat-label>Příjmení</mat-label>
+                <input [(ngModel)]="patientInfo.lastName" formControlName="lastName" matInput placeholder="Vaše příjmení" required>
+              </mat-form-field>
+
+              <mat-form-field>
+                <mat-label>Rodné číslo</mat-label>
+                <input [(ngModel)]="patientInfo.personalNumber" formControlName="personalNumber" matInput placeholder="Vaše rodné číslo" required>
+                <mat-hint>Rodné číslo ve formátu RRMMDD/ČČČ(Č)</mat-hint>
+              </mat-form-field>
+
+              <mat-form-field>
+                <mat-label>Zdravotní pojišťovna</mat-label>
+                <mat-select formControlName="insuranceCompany" [(ngModel)]="patientInfo.insuranceCompany">
+                  <mat-option *ngFor="let insuranceCompany of allInsuranceCompanies" [value]="insuranceCompany">
+                    {{insuranceCompany}}
+                  </mat-option>
+                </mat-select>
+              </mat-form-field>
+
+              <mat-form-field>
+                <mat-label>Telefonní číslo</mat-label>
+                <input [(ngModel)]="patientInfo.phoneNumber" formControlName="phoneNumber" matInput placeholder="Vaše telefonní číslo" required>
+                <mat-hint>Telefonní číslo ve formátu +420XXXXXXXXX</mat-hint>
+              </mat-form-field>
+              <mat-form-field>
+                <mat-label>E-mail</mat-label>
+                <input [(ngModel)]="patientInfo.email" formControlName="email" matInput placeholder="Váš e-mail" required>
+              </mat-form-field>
+            </form>
+          </mat-card-content>
+        </mat-card>
+
+        <div class="nav-buttons">
+          <div></div>
+          <button [disabled]="!canProceedToStep2"
+                  color="primary" mat-raised-button matStepperNext>
+            Další krok
+          </button>
+        </div>
+      </div>
+    </mat-step>
+    <mat-step label="Zdravotní informace" [completed]="canProceedToStep3">
+      <div class="container">
+        <mat-card class="py-0">
+          <mat-card-content>
+            <form>
+              <ng-container *ngFor="let q of questions; let i = index">
+                <div class="form-field">
+                  <mat-label>{{ q.label }}</mat-label>
+                  <mat-button-toggle-group [(ngModel)]="q.value" [name]="q.name">
+                    <mat-button-toggle value="true">Ano</mat-button-toggle>
+                    <mat-button-toggle value="false">Ne</mat-button-toggle>
+                  </mat-button-toggle-group>
+                </div>
+                <mat-divider *ngIf="i < questions.length - 1"></mat-divider>
+              </ng-container>
+            </form>
+          </mat-card-content>
+        </mat-card>
+        <div class="nav-buttons">
+          <button color="secondary" mat-raised-button matStepperPrevious>
+            Předchozí krok
+          </button>
+          <button [disabled]="!canProceedToStep3"
+                  color="primary" mat-raised-button matStepperNext>
+            Další krok
+          </button>
+        </div>
+      </div>
+    </mat-step>
+    <mat-step label="Odeslání požadavku">
+      <div class="container container--wide">
+        <div class="confirmation-step">
+          <mat-card *ngIf="basicInfoForm" class="py-0">
+            <mat-card-content>
+              <div class="form-field">
+                <mat-label>Jméno:</mat-label>
+                <mat-label *ngIf="basicInfoForm.controls['firstName']?.valid else unspecified;">
+                  {{ patientInfo.firstName }}
+                </mat-label>
+              </div>
+              <mat-divider></mat-divider>
+              <div class="form-field">
+                <mat-label>Příjmení:</mat-label>
+                <mat-label *ngIf="basicInfoForm.controls['lastName']?.valid else unspecified;">
+                  {{ patientInfo.lastName }}
+                </mat-label>
+              </div>
+              <mat-divider></mat-divider>
+              <div class="form-field">
+                <mat-label>Rodné číslo:</mat-label>
+                <mat-label *ngIf="basicInfoForm.controls['personalNumber']?.valid else unspecified;">
+                  {{ patientInfo.personalNumber }}
+                </mat-label>
+              </div>
+              <mat-divider></mat-divider>
+              <div class="form-field">
+                <mat-label>Pojištovna:</mat-label>
+                <mat-label *ngIf="basicInfoForm.controls['insuranceCompany']?.valid else unspecified;">
+                  {{ patientInfo.insuranceCompany }}
+                </mat-label>
+              </div>
+              <mat-divider></mat-divider>
+              <div class="form-field">
+                <mat-label>Telefonní číslo:</mat-label>
+                <mat-label *ngIf="basicInfoForm.controls['phoneNumber']?.valid else unspecified;">
+                  {{ patientInfo.phoneNumber }}
+                </mat-label>
+              </div>
+              <mat-divider></mat-divider>
+              <div class="form-field">
+                <mat-label>E-mail:</mat-label>
+                <mat-label *ngIf="basicInfoForm.controls['email']?.valid else unspecified;">
+                  {{ patientInfo.email }}
+                </mat-label>
+              </div>
+            </mat-card-content>
+          </mat-card>
+
+          <mat-card class="py-0">
+            <mat-card-content>
+              <ng-container *ngFor="let q of questions; let i = index">
+                <div class="form-field">
+                  <mat-label>{{ q.label }}</mat-label>
+                  <mat-label [ngSwitch]="q.value">
+                    <span *ngSwitchCase="'true'">Ano</span>
+                    <span *ngSwitchCase="'false'">Ne</span>
+                    <ng-container *ngSwitchDefault [ngTemplateOutlet]="unspecified"></ng-container>
+                  </mat-label>
+                </div>
+                <mat-divider *ngIf="i < questions.length - 1"></mat-divider>
+              </ng-container>
+            </mat-card-content>
+          </mat-card>
+        </div>
+
+        <div class="checkboxes">
+          <mat-checkbox [(ngModel)]="confirmationCheckboxValue">
+            Potvrzuji, že jsem nezamlčel/a žádné informace o svém zdravotním stavu.
+          </mat-checkbox>
+
+          <mat-checkbox [(ngModel)]="agreementCheckboxValue">
+            Souhlasím s aplikací očkovací látky proti onemocnění COVID-19.
+          </mat-checkbox>
+        </div>
+
+        <div class="nav-buttons">
+          <button color="secondary" mat-raised-button matStepperPrevious>
+            Předchozí krok
+          </button>
+          <button (click)="submit()"
+                  [disabled]="!canSubmit"
+                  color="primary" mat-raised-button matStepperNext>
+            Odeslat požadavek
+          </button>
+        </div>
+      </div>
+    </mat-step>
+  </mat-horizontal-stepper>
+</main>
+
+<footer>
+  <span>Powered by </span><a href="https://mild.blue" target="_blank">mild.blue</a>
+</footer>
+
+<ng-template #unspecified>
+  <span class="unspecified">
+    Neuvedeno
+  </span>
+</ng-template>
+
